requirements.txt
Flask>=3.0.0,<4.0.0
Flask-WTF>=1.2.1,<2.0.0
mysql-connector-python>=8.4.0,<9.0.0
passlib[bcrypt]>=1.7.4,<2.0.0
python-dotenv>=1.0.1,<2.0.0
fpdf2>=2.8.1,<3.0.0
gunicorn>=22.0.0,<23.0.0
pytest>=8.2.0,<9.0.0
pytest-cov>=5.0.0,<6.0.0

